Use SurfaceEffectInternal in Preview(roll forward)
Reason for revert: roll forward after fixing the test.

This CL is the same as aosp/2152154. The original CL was rolled back because a "DeferrableSurface garbage collected" failure in a seemingly unrelated test: `SurfaceEffectNodeTest#releaseNode_effectIsReleased`.
https://android-build.googleplex.com/builds/tests/view?invocationId=I43800010070419024&testResultId=TR89928138844967785

My theory is that we didn't close the Surface SurfaceOutput#invokedErrorListener_recreatePipeline that got attributed to SurfaceEffectNodeTest somehow. This CL makes sure to clean up DeferrableSurface(s) in both tests.

Bug: 238798655
Test: manual test and ./gradlew bOS
Change-Id: If0bab79ebd25c615c2f6a2a601546f217c997e7f