Fix EGLConfigAttributesJavaTest
Updated test to fix incorrect assumption
that the conversion of the HashSet to an
integer array would also maintain the same
order of the key/value pairs being added
to the HashSet managed internally by
EGLConfigAttributes. This is because
the behavior of entrySet does not return
the same ordering of entries across
different Android platform versions.

Fixes: 239165059
Test: re-ran EGLConfigAttibutesJavaTest on
API 17 emulator + API 33 device

Change-Id: I7f1fd4e46a55f69961df692ff824dc09e1c32af5