Remove Gallery
After discussion with Scott, we would remove the gallery section as it
does not add much value to the test app. It was meant to test navigation
while the cameras are active, which can already be achieved by switching between
ImageCapture and VideoCapture.

The focus of this app would be geared towards CameraX functionalities
instead.

Bug: 234080273
Test: Manual Testing on Pixel 5
Change-Id: I638f47040e2a86f15b963f2e663393c2cbbc39d1