Introduce tag
Add an unique tag as an identifier of EmbeddingRule.
The tag is optional, but must be unique if set. It allows to use
advanced Activity Embedding features. An example is to identify
the SplitRule in callbacks, such as SplitAttributesCalculator,
and it could be used as a key to update EmbeddingRules
via #registerRule. It throws an exception if there are
duplicated tags in static rules.

Test: ./gradlew window:window:ConnectedAndroidTest
Bug: 207494880
fixes: 240911858
Relnote: "
  - Add tag in EmbeddingRule
  - Add tag attributes
  - Add tag in all rule's Builder constructer
    - Add ActivityRule.Builder#setTag
    - Add SplitPairRule.Builder#setTag
    - Add SplitPlaceholderRule.Builder#setTag
"

Change-Id: I5869c6acb68d09e3f403979c8b53e519146f78ac