Moved the preferences okio factory to common common.
This allows clients to call the factory in common without having to create expect/actuals.

Best technique I could find that provides backward compatibility was creating a differently named factory method.  Also considered using @JvmName and changing the parameter name which allowed it to compile, but that caused issues in the kotlin with method ambiguity for existing code that didn't name the parameter.  Other possibility is fully new factory class with `create` method but different name.

Test: ./gradlew :datastore:datastore-preferences-core:check
Change-Id: I660ac9c62f999f40af1ea436a716034b74283964