Fix `testSwipe()`.
Updated `testSwipe()` into the same logic as `testFling()`, as they are
using the same internal logic in the UiAutomator.

The previous version of this test depends on the actual UI change to
assert the completion of the performed action, which may vary and thus
be unstable across different devices.

Fix: 240570957
Test: ./gradlew :test:uiautomator:integration-tests:testapp:cAT
Change-Id: I038a8f18a327c9eb4cea4574c9c2a0906a6b60d9