Hierarchical focus management
Add a composable to enable marking sub-trees of the composition as focus
enabled or focus disabled. There an implicit AND when there are nested,
they all have to be enabled for the inner sub-tree to be enabled.
This will be used on screen-managing components like swipe to dismiss or
horizontal pagers.

Add a composable function to query if the current point in the tree has
focus enabled.

Test: Added
Relnote: "Added facilities to mark parts of the tree as focus enabled,
and check the current status."

Change-Id: I827cb73ace07280bb0ad8677755efdf721c46147