Improve focused selection controls tests and add FocusedCheckboxSample
The tests are currently written by manually forcing focusability
even in non-touch-mode, and this doesn't work reliably in combination
with aosp/2108724.  It's a test-specific problem since non-touch-focus
works fine in a realistic scenario.

This changes the focus screenshot tests to use the more
mechanism requestInputMode(Keyboard), which is the same thing a
real-world non-touch-focus does.  It also adds a demo showing
use of requestInputMode to programmatically focus a checkbox
within the demo app.

Test: *ScreenshotTest#*_focus
Change-Id: Ibdd57e38524bde102ff411718d274d1076e3d5e8