Fix Core i18n tests failing because they expected PDT timezone
Relnote: Fix Core i18n tests failing because they expected PDT timezone

Some of the tests assumed that they run a device with the timezone set to PDT,
which was not always true.
Fixed by setting the default timezone to America/Los_Angeles in @Begin.

Also added special test to verify the interaction between the default timezone
and the timezone "captured" in a Calendar, pre and post N.

Bug: 241548422

Test: ./gradlew test connectedAndroidTest on all (emulator) API [19-33] images
Change-Id: I52f1e59f00493601228e631b4961a3269f60c486