Avoid ClassVerificationErrors in OnBackPressedDispatcher
Although we won't actually call any of the T specific OnBackInvoked APIs
on Android versions prior to T, the class loader does not know that. So
it attempts to load all of the classes and since theose APIs are not
present we get a ClassVerificationError.

We should move any initialization of these classes to static functions
that are specific to API 33.

RelNote: "Initializing a `OnBackPressedDispatcher` will not longer cause
`ClassVerificationError`s when using SDK versions prior to 33."
Test: verified with integration app and running test on 32- devices
Bug: 241739353

Change-Id: Ic32e1fb18a4f17692d4b62ddf9d37bb5e67c2771