Move startRemoteWorker to main thread
JavaDoc was promising it even before, but
in reality it was happening on ibinder thread
(if remote service is *actually* in another process).

Using ibinder thread for user callback is dangerous,
because one callback blocks all other messages
can't be delivered.

Test-wise: we use integration testapp, because
`RemoteListenableWorkerTest` runs in one process,
so ibinder thread isn't used: in the same process
IBinder is directly casted to ListenableWorkerImpl,
and methods are called on it.

fixes: 234211438
Test: integration test
Change-Id: I6c094ba008800ab2f0566121621b151c2063c3da