Fix LocalesLateOnCreateActivity to have correct values for defaultLocales and fix method culateApplicationLocales in AppCompatDelegateImpl
The method culateApplicationLocales was returning undefined locales for SDK 21 to SDK 23 when the input locale had both a language tag and a location tag. This was because the Locale.toString() method that was used returns the string representation with an underscore (for e.g. en_US), the expected string representation should contain a hyphen in the middle (for e.g. en-US). Hence to fix this issue, Locale.toLanguageTag() has been used instead.

Test: LocalesLateOnCreateTestCase.kt & go/forest-test-result-241547343

Bug: 243119645
Bug: 241547343
Change-Id: I5e7e41eada615667bb780d4976bda5e6344b7ff6