Allow InteractiveInstance to migrate between engines
If the watch face process gets killed, both WallpaperService and
WCS have mitigations. Its sometimes possible for these to clash,
specifically if WCS gets in first it'll reopen an
InteractiveInstance. WallpaperService will try to restart the
watchface and this leads to a call to atach which detaches from
the old engine. Unfortunately WCS doesn't (and arguably shouldn't)
know about this so it carries on sending commands to a defunct
instance leading to a lot of logcat spam.

This patch adds a mitigation to this scenario by transparently
migrating the InteractiveInstance from the old engine to the new
one.

Bug: 242878416
Test: Added new tests. Was unable to reproduce root cause bug, we're reasonably confident this fixes it however!
Change-Id: I0826dd7a0d7cf8507e7b3810caee739c2ef57bf8