Save the states only for the currently visible items of lazy layouts
When we save the state of the screen in cases like navigation to other screen or activity rotation we were saving the state for all items which were ever visible in the lazy layouts like LazyColumn. We want to save state while user is scrolling so they not lose it when they scroll back, however when this screen is not active anymore keeping the states for all items is not efficient as we easily can fill up the whole available limit provided by the Bundle. Instead it could be reasonable to only save states for the items which were visible when the screen state save happens.

Relnote: N/A
Fixes: 240837359
Fixes: 242589959
Fixes: 212982463
Test: added LazyLayoutStateRestorationTest. Removed list and grid specific tests as this logic is specific to LazyLayout and should be tested there
Change-Id: Ic04e855cdaa8d25846c55b5671afe7e8e7337b13