Have FlowLiveData.asFlow use callbackFlow
Now that awaitClose() is no longer experimental,
we should have FlowLiveData use callbackFlow, rather
than just flow to utilize its safety metrics.

The reordering in LiveDataAsFlowTest.reusingFlow was
done to match the order in which things are done in
the other tests (launching testScope before mainScope).
The original order of this test was not allowing scopes.triggerAllActions() to catch the closing of
both coroutines that were launched, so we separated them.

RelNote: "`FlowLiveData.asFlow()` now creates a `callbackFlow`
rather than using its own `Channel` implementation to ensure
thread-safety and context preservation."
Test: ensure all tests in LiveDataAsFlowTest pass
Bug: 200596935
Change-Id: I4a8b270a92ac3779cbb3d56dae585bf6876396d2