Do not track capture metadata in Packet<>
This CL walks back from one of the decision made early.

In the previous CL, I tracked the metadata so that in the  Yuv->Jpeg case, we can postpone writing the Exif until the final step, which will result in better performance.

Howevewr after inspecting the logic in Yuv2RgbProcessor, I realized that the overhead of reading/writing the metadata is trivial. The optimization is premature because it adds a non-trival complexity to our code.

Bug: 240998057
Test: manual test and ./gradlew bOS
Change-Id: I9b06138c1056cdad57e51ddd2fec5e3a45a0e8cc