Post processNextRequest to main thread
There are 2 cases could call
ImageCaptureRequestProcessor#processNextRequest in
the background thread, and which causes the thread
checking failure in the ImageCapture#submitStillCaptureRequest.

Case 1) The ImageCaptureRequestProcessor#onImageClose
is signaled in the background thread.
Case 2) The result handling of the ImageCaptor#capture
uses the directExecutor, which may be called in the
background thread.

This CL makes a quick fix by posting the processNextRequest
to the main thread.
The issue may not happen after migrating to the
TakePictureManager.

Bug: 245238082
Test: ImageCaptureTest
Change-Id: Ib9267daeb00bc1b37329b8e7ed667ea0c73e90b7