Make Packet#getExif() nullable. (again)
The Exif needs to be nullable, because we can't create an Exif instance from a YUV image. For YUV image, we can extract the exif into a ExifData instance. Unfortunately Exif and ExifData are not compatible, e.g. we can't convert one to another. We might end up tracking both ExifData and Exif in Packet.

Bug: 240998057
Test: manual test and ./gradlew bOS
Change-Id: I6a8f0d0f0b9a077c62c9e5a45d5cdafd118251f3