[Extensions] Revert onDisableSession sleeping
The CL reverts aosp/2207335 due to concern of blocking
camera thread.

The issue no longer happens after the change in
LifecycleStatusChangeStressTest even without the onDisableSession
sleeping.  The new change in LifecycleStatusChangeStressTest
waits until preview is streaming and then stop the camera in each
repeating block, which increase the duration between onEnableSession
and onDisableSession.

This is more likely the normal scenarios user runs into.
Previous flow is like
  repeating(10) {
     moveToState(Lifecycle.State.CREATED)
     moveToState(Lifecycle.State.RESUMED)
  }

which is very extreme and uncommon.  Issues detected by
this extreme flow probably are not met by regular users.

Bug: 245226085
Test: LifecycleStatusChangeStressTest
Change-Id: If41cdc2144db409c70a87daefd13f39ae23ea402