Moved IntentExt.kt to utils package.
Turned out the file doesn't compile when imported to G3 - AbstractMessageLite is unresolved. There is also no clear way to update the copybara script, so it could replace the import - the package is the same and so there is no import line. Moved to the utils package so there is an explicit import line, and so the copybara script could be updated later.

Test: IntentExtTest
Bug: 232379313
Change-Id: I73fb26b9f34866b1a36c4ebc5a6735c62473890b